+++ q01_10/003.py
@@ -5,7 +5,7 @@ import lib0110
 if __name__ == '__main__':
     img = cv2.imread("Gasyori100knock/Question_01_10/imori.jpg")
     grayimage = lib0110.BGR2GRAY(img)
-    ans = lib0110.binalization(grayimage,128)
+    ans = lib0110.binalization(grayimage,128).astype(np.uint8)
 
     cv2.imshow("imori", ans)
     cv2.waitKey(0)
