+++ q01_10/006.py
@@ -4,7 +4,7 @@ import lib0110
 
 if __name__ == '__main__':
     img = cv2.imread("Gasyori100knock/Question_01_10/imori.jpg")
-    ans = lib0110.color_reduction(img)
+    ans = lib0110.color_reduction(img).astype(np.uint8)
     
     cv2.imshow("imori", ans)
     cv2.waitKey(0)
